Optimize IVR call retries tasks
Call tasks for busy and failed calls have been merged into a single
task. Added a db-index for failed calls because this task is going to be
executed at regular intervals.
Updated tests to verify that we release the active call line on
throttled channels when call ends up in busy/no-answer state.